requirements.txt
flask==3.0.0
numpy==1.26.0
gunicorn==22.0.0
Flask-Cors==4.0.1
pandas==2.2.2
google-generativeai==0.5.4
python-dotenv==1.0.1